requirements.txt
flask==2.3.3
pillow==10.0.1
python-multipart==0.0.6